v.1.0 VideoStore and MediaRevCo Integration
See parts b and c of the workflow:

The publish/subscribe between reviewCo and VideoStore is now done using
msmq. Messages for subscriptions to reviews on media item creation will
arrive in if VideoStore is up and reviewCo is down or the VideoStore
started before reviewCo. Same with saving a review on reviewCo the
review will arrive to videoStore when it is up regardless of its status
at the time of saving the message.